(E) April 17 and 18 (midnight), 2025
This is will be the last update for the first application replication process (Notes app). Expect within this coming days for the 2nd type of Android application, different functons and UI/UX implementation.

You can check the screenshots and even the screen recording that I've uploaded to see the overall result/ functionality so far.